btId 외래키 문제 해결해야함

+++ src/main/java/platform/game/service/entity/Battle.java
@@ -25,7 +25,7 @@ public class Battle {
     
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int btId;
+    private long btId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "btHostMemId", referencedColumnName = "memId")
